browser: refactor data format, bugfixes; emms: add emms-track-p
Subitems are now stored in a tree of "bdata" objects, which is
generated when initially displaying the buffer. This makes rendering
simpler and also fixes some bugs where tracks were not being sorted
correctly in the browser and/or the playlist.

Adding items to the playlist now inserts 'group' names when you add a
whole album, artist, etc at a time. These names will be thrown away if
you run emms-shuffle. The sorting routines will correctly sort the
buffer, throwing away the group tags, but then throw an error because
the buffer is not the size they expected it to be. Other playlist
manipulation routines like next/previous should skip over the group
names - if they don't, it's a bug in the playlist code.

An example of the new interface is at:
http://repose.cx/dump/emms-browser.png

darcs-hash:20060614160048-4e3e3-82a8a0d1678b0a2d9fcfc6ca385d5b56963aedbe.gz