Distinguish `emms-track-description' and `emms-track-force-description'
The idea is that emms-track-force-description is only allowed to be
used when inserting descriptions into a playlist buffer.  This lets us
distinguish between the content and layout of the text, for add-ons
like emms-mark that want to do something decorative with the text.

darcs-hash:20070129091330-1bfb2-efe486ed3f4b256cf5817a1b36d66448a6c14196.gz