use-existing-functionality-for-renaming-files.dpatch
When I implemented the renaming stuff, I was not aware that Ye already thought
of this feature.  When editing a track, one can edit the filename and the file
will be renamed and the cache updated.  Now renaming by hitting `R'
(`emms-tag-editor-rename') in the playlist makes use of this functionality
instead of doing it on its own.

darcs-hash:20070817100310-c06f4-49ce1f24221f5701fd77be87f4823cf594bb1146.gz