Add support for composer and performer tags.
The browser keybindings are somewhat discontiguous to avoid breaking
people's finger macros: browsing by composer and performer are `b 5' and
`b 6' respectively, where artist is `b 1'.

In the tag editor, I changed the note format character to %;, making room
for composer as %c.

I've not added support to last.fm for composer or performer, as it doesn't
seem to support either. (I suspect that in most cases the composer should
be reported as the artist if it is present, but that is not impolemented
here.)

darcs-hash:20070818133412-984cd-07ed55c9bf357ee6119f6d70fc2f267e9e8ecef9.gz