emms-playlist-mode: Use current buffer for clear and kill.
In this case of the clear and kill commands, it makes the most sense
to operate on the current buffer, rather than the currently-designated
EMMS playlist buffer.

Additionally, fix a bug when doing "a" or "RET" on something that
isn't the current EMMS track.

darcs-hash:20080130064148-1bfb2-0cf2b8b59d0f338d67cd21e272bb305bd9b9ada1.gz