* emms-browser.el (emms-browser-make-hash-by): Check that field is present before attempting to add entries.
Addresses an issue'm encountered when using lastfm integration and mpd
as the player.

After requesting streaming from last.fm, a new track gets added to
EMMS' database, which is named after the streaming URL but has no
other fields.  Thus, when adding entries or making searches, checking
the the field at hand is non-nil is needed.  This patch checks that
for emms-browser-make-hash-by.