Make cons cell usable when replacing a tag
As reported on the mailing list[1][2], it's not possible to use a cons
cell like '\,(downcase \1)' in emms-tag-editor-replace-in-tag.

After looking at how query-replace-regexp handle this case, I patch
'emms-tag-editor-replace-in-tag'.

I also fix a bug when using a catch all regexp like  '\(.*\)', the
generator will first try to match on the current line, to replace
multiple occurrences in the same string.

But '(re-search-forward "\\(.*\\)" (line-end-position) t)' will match
the end-of-line, making the replacement looping on the same line.

* lisp/emms-tag-editor.el (emms-tag-editor-replace-in-tag): Prevent
  looping on the same line by matching end-of-line when at the end of the
  line.
  Make the "map-y-or-n-p" generator returning the matched string and the
  calculated replacement and adapt the prompter and the actor
  accordingly.

Footnotes:
[1]  http://lists.gnu.org/archive/html/emms-help/2005-07/msg00010.html

[2]  https://lists.gnu.org/archive/html/emms-help/2012-01/msg00002.html