Clean: emms-track-file-p replaces eq file type checks
There were about 10 checks throughout the code checking if a track is a
file type. File type tracks are specific enough for a predicate function.

Add: EMMS warning if no tagging program for audio file type

Warn the user when there is no tag writing program available to EMMS
for modifying the actual audio files. Current behavior allows the tags
to be edited, saved to the EMMS cache db, and then silently skips past
the part where the metadata would be applied to the file itself.

The changes here prevent the user from even opening the editor for a
file that has no tag writing program configured. This can be
overridden with the prefix arg when user wants to edit the tags
anyway, knowing they will only be saved in the emms-cache-db and not
written to the source audio files.

Patch by Grant Shangreaux