* emms-browser.el: Fix bug in emms-browser-format-spec
Sometimes, emms-browser-format-spec will throw an "Invalid format
character" error, even when supplied a valid character format. This
happens because the propertized string returned by
emms-browser-format-line may contain an item with a car but no cdr
(e.g.: ("y") instead of ("y" . 2023)). This patch fixes that by
throwing the error only if the propertized string doesn't contain the
key. Then, in circumstances where it does contain the key but the
value is nil, it will supply an empty string to the text formatting
procedure.

Patch by Scarlett McAllister