Remove most length limits from Vorbis bindat specs
Many of the field length limits were arbitrarily chosen and even too
strict in practice.  It is better to check against the length of input
data, which forms trivially an upper bound for the length of any
field.

Do still keep maximum metadata peek sizes to limit reading of
excessive amounts of data, but increase the limit to 16 MB.